+++ Cargo.toml
@@ -12,10 +12,9 @@ A library to parse multiple MRT files in parallel and creating timestamp-sorted 
 keywords = ["bgp", "bgpkit", "mrt"]
 
 [dependencies]
-bgpkit-broker = "0.3"
-bgpkit-parser = {git="https://github.com/bgpkit/bgpkit-parser.git", branch="main"}
-spmc = "0.3.0"
+bgpkit-broker = "0.4.0"
+bgpkit-parser = "0.7.3"
 itertools="0.10.3"
 rayon = "1.5.1"
 tracing = "0.1"
-tracing-subscriber = "0.3.3"
+tracing-subscriber = "0.3"
